+++ src/main/java/org/mmarini/wheelly/apps/Report.java
@@ -68,11 +68,13 @@ public class Report {
             ReportProcess.maxReport("trainingLayers.move.values"), // max probability for move
             ReportProcess.gmReport("trainingLayers.move.values"), // gm probability for move
             ReportProcess.maxMinRatioReport("trainingLayers.move.values"), // max/min probability ratio for move
+            ReportProcess.saturationReport("trainingLayers.move[1].values"), // saturation ratio for move
             ReportProcess.maxGMRatioReport("trainingLayers.move.values"), // max/mean probability for move
 
             ReportProcess.statsReport("trainingLayers.sensorAction.values"), // max probability for move
             ReportProcess.maxReport("trainingLayers.sensorAction.values"), // max probability for sensor
             ReportProcess.gmReport("trainingLayers.sensorAction.values"), // gm probability for sensor
+            ReportProcess.saturationReport("trainingLayers.sensorAction[1].values"), // saturation for sensor
             ReportProcess.maxMinRatioReport("trainingLayers.sensorAction.values"), // max probability for sensor
             ReportProcess.maxGMRatioReport("trainingLayers.sensorAction.values"),