+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue [#500](https://github.com/m-marini/wheellyj/issues/500): Add policy stats in report
 - Issue [#502](https://github.com/m-marini/wheellyj/issues/502): Add relocate robot button
 - Issue [#506](https://github.com/m-marini/wheellyj/issues/506): Add move to goal
+- Issue [#512](https://github.com/m-marini/wheellyj/issues/512):  Add saturation output level
 
 ### Fixed